requirements.txt
asgiref
numpy
async-timeout
asyncio
certifi
charset-normalizer
coreapi
coreschema
Django
django-filter
django-rest-swagger
djangorestframework
djangorestframework-simplejwt
docopt
drf-yasg
grpcio
grpcio-tools
idna
inflection
itypes
Jinja2
lark-parser
MarkupSafe
openapi-codec
packaging
proto-parser
protobuf
psycopg2
PyJWT
pytz
redis
requests
ruamel.yaml
simplejson
sqlparse
TornadIO2
tornado
tornado-redis
typing
tzdata
uritemplate
urllib3
websockets
